daemon: subclassing the connection class separation of connectionpool\connection subclasses
