+++ spruned/daemon/p2p/p2p_connection.py
@@ -1,6 +1,4 @@
 import asyncio
-import random
-from typing import List
 
 import async_timeout
 from pycoin.message.InvItem import InvItem, ITEM_TYPE_TX, ITEM_TYPE_BLOCK, ITEM_TYPE_MERKLEBLOCK
@@ -10,15 +8,52 @@ from pycoinnet.networks import MAINNET
 from pycoinnet.inv_batcher import InvBatcher
 from pycoinnet.version import version_data_for_peer
 from spruned.application.logging_factory import Logger
-from spruned.daemon.abstracts import ConnectionPoolAbstract
-from spruned.daemon.p2p.utils import dns_bootstrap_servers
+from spruned.application.tools import check_internet_connection, async_delayed_task
+from spruned.daemon.abstracts import ConnectionAbstract
+from spruned.daemon.connectionpool_base_impl import BaseConnectionPool
 
 
-class P2PConnection:
+class P2PConnection(ConnectionAbstract):
+    @property
+    def start_score(self):
+        raise NotImplementedError
+
+    @property
+    def hostname(self):
+        raise NotImplementedError
+
+    @property
+    def score(self):
+        raise NotImplementedError
+
+    @property
+    def version(self):
+        raise NotImplementedError
+
+    @property
+    def last_header(self):
+        raise NotImplementedError
+
+    def is_online(self) -> bool:
+        raise NotImplementedError
+
+    def ping(self, timeout=None):
+        raise NotImplementedError
+
+    @property
+    def errors(self):
+        raise NotImplementedError
+
+    def subscribe(self, channel, on_subscription, on_traffic):
+        raise NotImplementedError
+
+    def add_error(self, *a):
+        raise NotImplementedError
+
     def __init__(
             self, host, port, peer=Peer, network=MAINNET, loop=asyncio.get_event_loop(),
             peer_event_factory=PeerEvent):
-        self.host = host
+        self.hostname = host
         self.port = port
         self._peer_factory = peer
         self._peer_network = network
@@ -67,21 +102,32 @@ class P2PConnection:
         return bool(self.peer)
 
     async def connect(self):
-        self._reader, self._writer = await asyncio.open_connection(host=self.host, port=self.port)
-        peer = Peer(
-            self._reader,
-            self._writer,
-            self._peer_network.magic_header,
-            self._peer_network.parse_from_data,
-            self._peer_network.pack_from_data
-        )
-        version_data = version_data_for_peer(peer)
-        peer.version = await peer.perform_handshake(**version_data)
+        try:
+            with async_timeout.timeout(self._timeout):
+                self._reader, self._writer = await asyncio.open_connection(host=self.hostname, port=self.port)
+                peer = Peer(
+                    self._reader,
+                    self._writer,
+                    self._peer_network.magic_header,
+                    self._peer_network.parse_from_data,
+                    self._peer_network.pack_from_data
+                )
+                version_data = version_data_for_peer(peer)
+                peer.version = await peer.perform_handshake(**version_data)
+        except Exception as e:
+            Logger.p2p.error('Exception connecting to %s (%s)', self.hostname, e)
+            await self.on_error('connect')
+            return
+
         self._version = peer.version
         self.peer = peer
         self._event_handler = self._peer_event_factory(self.peer)
+        await self.on_connect()
         self._setup_events_handler()
-        return self
+
+    async def on_connect(self):
+        for callback in self._on_connect_callbacks:
+            self.loop.create_task(callback(self))
 
     async def disconnect(self):
         try:
@@ -129,33 +175,28 @@ class P2PConnection:
                 Logger.p2p.error('Error InvType: %s, %s, %s', event_handler, name, item)
 
 
-class P2PConnectionPool(ConnectionPoolAbstract):
-    @property
-    def established_connections(self) -> List:
-        raise NotImplementedError
-
-    def add_on_connected_observer(self, observer):
-        raise NotImplementedError
-
-    def add_header_observer(self, observer):
-        raise NotImplementedError
-
-    def is_online(self) -> bool:
+class P2PConnectionPool(BaseConnectionPool):
+    def on_peer_received_peers(self, peer: ConnectionAbstract):
         raise NotImplementedError
 
     def __init__(
-            self, inv_batcher=InvBatcher, network=MAINNET, connections=1,
+            self,
+            peers=list(),
+            network_checker=check_internet_connection,
+            delayer=async_delayed_task,
             loop=asyncio.get_event_loop(),
-            network_checker=None, peers_factory=dns_bootstrap_servers
+            use_tor=False,
+            connections=3,
+            sleep_no_internet=30,
+            inv_batcher=InvBatcher,
+            network=MAINNET,
     ):
-        self._connections = []
-        self._peers = []
+        super().__init__(
+            peers=peers, network_checker=network_checker, delayer=delayer,
+            loop=loop, use_tor=use_tor, connections=connections, sleep_no_internet=sleep_no_internet
+        )
         self._batcher_factory = inv_batcher
         self._network = network
-        self._required_connections = connections
-        self.loop = loop
-        self._peers_factory = peers_factory
-        self._network_checker = network_checker
 
     @property
     def available(self):
@@ -175,13 +216,8 @@ class P2PConnectionPool(ConnectionPoolAbstract):
             await batcher.add_peer(peer_event_handler)
         return batcher
 
-    async def bootstrap_peers(self):
-        with async_timeout.timeout(10):
-            self._peers = await self._peers_factory(self._network)
-
     async def connect(self):
         while not self._peers:
-            self.loop.create_task(self.bootstrap_peers())
             await asyncio.sleep(5)
 
         while 1:
@@ -211,16 +247,6 @@ class P2PConnectionPool(ConnectionPoolAbstract):
             connection.add_on_error_callback(self.on_peer_error)
         else:
             Logger.p2p.warning('Connection to %s - %s failed', host, port)
-
-    def _pick_peer(self):
-        if not self._peers:
-            return
-        return random.choice(self._peers)
-
-    def _pick_connection(self):
-        if not self._connections:
-            return
-        return random.choice(self.connections)
 
     async def get(self, inv_item: InvItem):
         batcher = await self._get_batcher()
